thoas jefferson

+++ git-all/uninstall.sh
@@ -3,23 +3,43 @@
 # UNINSTALLER 
 
 # ask for sudo permissions
-[ "$UID" -eq 0 ] || exec sudo bash "$0" "$@"
 
-user_dir="$(eval echo ~$SUDO_USER)"
+# ROOT PERMISSIONS
+
+if [ ! "$UID" -eq 0 ] 
+then
+    echo -e "FATAL ERROR:"
+    echo -e "   Can't uninstall without root permissions."
+    echo -e "   Please run 'sudo ./$(basename "$0")'."
+    exit 1
+fi
 
 
-echo -e "DELETING .DESKTOP FILE..." 
-sudo rm /usr/share/applications/git-all.desktop
+
+os="$(uname)"
+
+
+
 
 echo -e "DELETING CONFIG FOLDER..."
-rm -rf "$user_dir/.config/ez-scripts/git-all"
+rm -rf "$HOME/.config/ez-scripts/git-all"
 
 # if scripts folder is empty, remove it too
-[ "$(ls -A "$user_dir/.config/ez-scripts/")" ] || rmdir "$user_dir/.config/ez-scripts/"
+[ "$(ls -A "$HOME/.config/ez-scripts/")" ] || rmdir "$HOME/.config/ez-scripts/"
 
 
 echo -e "DELETING EXECUTABLE..." 
 rm "/usr/bin/git-all"
+
+
+
+if [ ! "$os" == "Darwin" ] 
+then
+
+echo -e "DELETING .DESKTOP FILE..." 
+sudo rm /usr/share/applications/git-all.desktop
+
+fi
 
 echo -e "Thank you for using git-all. -Dmytro"
